+++ lib/identify.c
@@ -62,21 +62,21 @@ static int line_intersect(const struct quirc_point *p0,
 	return 1;
 }
 
-static void perspective_setup(double *c,
+static void perspective_setup(float *c,
 			      const struct quirc_point *rect,
-			      double w, double h)
+			      float w, float h)
 {
-	double x0 = rect[0].x;
-	double y0 = rect[0].y;
-	double x1 = rect[1].x;
-	double y1 = rect[1].y;
-	double x2 = rect[2].x;
-	double y2 = rect[2].y;
-	double x3 = rect[3].x;
-	double y3 = rect[3].y;
+	float x0 = rect[0].x;
+	float y0 = rect[0].y;
+	float x1 = rect[1].x;
+	float y1 = rect[1].y;
+	float x2 = rect[2].x;
+	float y2 = rect[2].y;
+	float x3 = rect[3].x;
+	float y3 = rect[3].y;
 
-	double wden = w * (x2*y3 - x3*y2 + (x3-x2)*y1 + x1*(y2-y3));
-	double hden = h * (x2*y3 + x1*(y2-y3) - x3*y2 + (x3-x2)*y1);
+	float wden = w * (x2*y3 - x3*y2 + (x3-x2)*y1 + x1*(y2-y3));
+	float hden = h * (x2*y3 + x1*(y2-y3) - x3*y2 + (x3-x2)*y1);
 
 	c[0] = (x1*(x2*y3-x3*y2) + x0*(-x2*y3+x3*y2+(x2-x3)*y1) +
 		x1*(x3-x2)*y0) / wden;
@@ -93,24 +93,24 @@ static void perspective_setup(double *c,
 		hden;
 }
 
-static void perspective_map(const double *c,
-			    double u, double v, struct quirc_point *ret)
+static void perspective_map(const float *c,
+			    float u, float v, struct quirc_point *ret)
 {
-	double den = c[6]*u + c[7]*v + 1.0;
-	double x = (c[0]*u + c[1]*v + c[2]) / den;
-	double y = (c[3]*u + c[4]*v + c[5]) / den;
+	float den = c[6]*u + c[7]*v + 1.0;
+	float x = (c[0]*u + c[1]*v + c[2]) / den;
+	float y = (c[3]*u + c[4]*v + c[5]) / den;
 
 	ret->x = (int) rint(x);
 	ret->y = (int) rint(y);
 }
 
-static void perspective_unmap(const double *c,
+static void perspective_unmap(const float *c,
 			      const struct quirc_point *in,
-			      double *u, double *v)
+			      float *u, float *v)
 {
-	double x = in->x;
-	double y = in->y;
-	double den = -c[0]*c[7]*y + c[1]*c[6]*y + (c[3]*c[7]-c[4]*c[6])*x +
+	float x = in->x;
+	float y = in->y;
+	float den = -c[0]*c[7]*y + c[1]*c[6]*y + (c[3]*c[7]-c[4]*c[6])*x +
 		c[0]*c[4] - c[1]*c[3];
 
 	*u = -(c[1]*(y-c[5]) - c[2]*c[7]*y + (c[5]*c[7]-c[4])*x + c[2]*c[4]) /
@@ -299,16 +299,16 @@ static uint8_t otsu(const struct quirc *q)
 	}
 
 	// Calculate weighted sum of histogram values
-	double sum = 0;
+	float sum = 0;
 	unsigned int i = 0;
 	for (i = 0; i <= UINT8_MAX; ++i) {
 		sum += i * histogram[i];
 	}
 
 	// Compute threshold
-	double sumB = 0;
+	float sumB = 0;
 	unsigned int q1 = 0;
-	double max = 0;
+	float max = 0;
 	uint8_t threshold = 0;
 	for (i = 0; i <= UINT8_MAX; ++i) {
 		// Weighted background
@@ -322,10 +322,10 @@ static uint8_t otsu(const struct quirc *q)
 			break;
 
 		sumB += i * histogram[i];
-		const double m1 = sumB / q1;
-		const double m2 = (sum - sumB) / q2;
-		const double m1m2 = m1 - m2;
-		const double variance = m1m2 * m1m2 * q1 * q2;
+		const float m1 = sumB / q1;
+		const float m2 = (sum - sumB) / q2;
+		const float m1m2 = m1 - m2;
+		const float variance = m1m2 * m1m2 * q1 * q2;
 		if (variance >= max) {
 			threshold = i;
 			max = variance;
@@ -582,7 +582,7 @@ static void find_alignment_pattern(struct quirc *q, int index)
 	int size_estimate;
 	int step_size = 1;
 	int dir = 0;
-	double u, v;
+	float u, v;
 
 	/* Grab our previous estimate of the alignment pattern corner */
 	memcpy(&b, &qr->align, sizeof(b));
@@ -648,10 +648,10 @@ static void find_leftmost_to_line(void *user_data, int y, int left, int right)
 	}
 }
 
-static double length(struct quirc_point a, struct quirc_point b)
+static float length(struct quirc_point a, struct quirc_point b)
 {
-	double x = abs(a.x - b.x) + 1;
-	double y = abs(a.y - b.y) + 1;
+	float x = abs(a.x - b.x) + 1;
+	float y = abs(a.y - b.y) + 1;
 	return sqrt(x * x +  y * y);
 }
 /* Estimate grid size by determing distance between capstones
@@ -664,15 +664,15 @@ static void measure_grid_size(struct quirc *q, int index)
 	struct quirc_capstone *b = &(q->capstones[qr->caps[1]]);
 	struct quirc_capstone *c = &(q->capstones[qr->caps[2]]);
 
-	double ab = length(b->corners[0], a->corners[3]);
-	double capstone_ab_size = (length(b->corners[0], b->corners[3]) + length(a->corners[0], a->corners[3]))/2.0;
-	double ver_grid = 7.0 * ab / capstone_ab_size;
+	float ab = length(b->corners[0], a->corners[3]);
+	float capstone_ab_size = (length(b->corners[0], b->corners[3]) + length(a->corners[0], a->corners[3]))/2.0;
+	float ver_grid = 7.0 * ab / capstone_ab_size;
 
-	double bc = length(b->corners[0], c->corners[1]);
-	double capstone_bc_size = (length(b->corners[0], b->corners[1]) + length(c->corners[0], c->corners[1]))/2.0;
-	double hor_grid = 7.0 * bc / capstone_bc_size;
+	float bc = length(b->corners[0], c->corners[1]);
+	float capstone_bc_size = (length(b->corners[0], b->corners[1]) + length(c->corners[0], c->corners[1]))/2.0;
+	float hor_grid = 7.0 * bc / capstone_bc_size;
 	
-	double grid_size_estimate = (ver_grid + hor_grid) / 2;
+	float grid_size_estimate = (ver_grid + hor_grid) / 2;
 
 	int ver = (int)((grid_size_estimate - 17.0 + 2.0) / 4.0);
 	
@@ -703,7 +703,7 @@ static int fitness_cell(const struct quirc *q, int index, int x, int y)
 
 	for (v = 0; v < 3; v++)
 		for (u = 0; u < 3; u++) {
-			static const double offsets[] = {0.3, 0.5, 0.7};
+			static const float offsets[] = {0.3, 0.5, 0.7};
 			struct quirc_point p;
 
 			perspective_map(qr->c, x + offsets[u],
@@ -806,7 +806,7 @@ static void jiggle_perspective(struct quirc *q, int index)
 	struct quirc_grid *qr = &q->grids[index];
 	int best = fitness_all(q, index);
 	int pass;
-	double adjustments[8];
+	float adjustments[8];
 	int i;
 
 	for (i = 0; i < 8; i++)
@@ -816,9 +816,9 @@ static void jiggle_perspective(struct quirc *q, int index)
 		for (i = 0; i < 16; i++) {
 			int j = i >> 1;
 			int test;
-			double old = qr->c[j];
-			double step = adjustments[j];
-			double new;
+			float old = qr->c[j];
+			float step = adjustments[j];
+			float new;
 
 			if (i & 1)
 				new = old + step;
@@ -998,7 +998,7 @@ fail:
 
 struct neighbour {
 	int		index;
-	double		distance;
+	float		distance;
 };
 
 struct neighbour_list {
@@ -1015,7 +1015,7 @@ static void test_neighbours(struct quirc *q, int i,
 		const struct neighbour *hn = &hlist->n[j];
 		for (int k = 0; k < vlist->count; k++) {
 			const struct neighbour *vn = &vlist->n[k];
-			double squareness = fabs(1.0 - hn->distance / vn->distance);
+			float squareness = fabs(1.0 - hn->distance / vn->distance);
 			if (squareness < 0.2)
 				record_qr_grid(q, hn->index, i, vn->index);
 		}
@@ -1037,7 +1037,7 @@ static void test_grouping(struct quirc *q, unsigned int i)
 	 */
 	for (j = 0; j < q->num_capstones; j++) {
 		struct quirc_capstone *c2 = &q->capstones[j];
-		double u, v;
+		float u, v;
 
 		if (i == j)
 			continue;
